chore: change assets func to vuetifyxjs
